Reorganize servers, cmd files are now very barebones, add cmux so /ok:80 works and AVAILABLE is shown in CCIS, reorganize wirepod
