+++ chipper/cmd/vosk/main.go
@@ -1,75 +1,10 @@
 package main
 
 import (
-	"fmt"
-
-	"github.com/kercre123/chipper/pkg/jdocsserver"
-	sdkWeb "github.com/kercre123/chipper/pkg/sdkapp"
-	"github.com/kercre123/chipper/pkg/tokenserver"
-	wp "github.com/kercre123/chipper/pkg/wirepod-prs"
-	wpweb "github.com/kercre123/chipper/pkg/wirepod-ws"
-
-	// the import path should be the only thing you need to change if you want another stt engine
-	stt "github.com/kercre123/chipper/pkg/wirepod-stt/vosk"
-
-	pb "github.com/digital-dream-labs/api/go/chipperpb"
-	"github.com/digital-dream-labs/api/go/jdocspb"
-	"github.com/digital-dream-labs/api/go/tokenpb"
-	"github.com/kercre123/chipper/pkg/server"
-
-	//	grpclog "github.com/digital-dream-labs/hugh/grpc/interceptors/logger"
-
-	grpcserver "github.com/digital-dream-labs/hugh/grpc/server"
-	"github.com/digital-dream-labs/hugh/log"
+	"github.com/kercre123/chipper/pkg/initwirepod"
+	stt "github.com/kercre123/chipper/pkg/wirepod/stt/vosk"
 )
 
 func main() {
-	log.SetJSONFormat("2006-01-02 15:04:05")
-	startServer()
-}
-
-func startServer() {
-	srv, err := grpcserver.New(
-		grpcserver.WithViper(),
-		//grpcserver.WithLogger(log.Base()),
-		grpcserver.WithReflectionService(),
-
-		grpcserver.WithUnaryServerInterceptors(
-		//			grpclog.UnaryServerInterceptor(),
-		),
-
-		grpcserver.WithStreamServerInterceptors(
-		//			grpclog.StreamServerInterceptor(),
-		),
-	)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	p, err := wp.New(stt.Init, stt.STT, stt.Name)
-	go wpweb.StartWebServer()
-	wp.InitHoundify()
-	go sdkWeb.BeginServer()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	s, _ := server.New(
-		//server.WithLogger(logger.Base()),
-		server.WithIntentProcessor(p),
-		server.WithKnowledgeGraphProcessor(p),
-		server.WithIntentGraphProcessor(p),
-	)
-
-	tokenServer := tokenserver.NewTokenServer()
-	jdocsServer := jdocsserver.NewJdocsServer()
-	//jdocsserver.IniToJson()
-
-	pb.RegisterChipperGrpcServer(srv.Transport(), s)
-	jdocspb.RegisterJdocsServer(srv.Transport(), jdocsServer)
-	tokenpb.RegisterTokenServer(srv.Transport(), tokenServer)
-
-	srv.Start()
-	fmt.Println("\033[33m\033[1mServer started successfully!\033[0m")
-	<-srv.Notify(grpcserver.Stopped)
+	initwirepod.StartServer(stt.Init, stt.STT, stt.Name)
 }
