+++ chipper/pkg/logger/logger.go
@@ -5,8 +5,18 @@ import (
 	"os"
 )
 
-func Println(a ...any) {
+var debugLogging bool = true
+
+func Init() {
 	if os.Getenv("DEBUG_LOGGING") == "true" {
+		debugLogging = true
+	} else {
+		debugLogging = false
+	}
+}
+
+func Println(a ...any) {
+	if debugLogging {
 		fmt.Println(a...)
 	}
 }
