+++ chipper/pkg/wirepod/preqs/intent.go
@@ -1,0 +1,32 @@
+package processreqs
+
+import (
+	"strconv"
+
+	"github.com/kercre123/chipper/pkg/logger"
+	"github.com/kercre123/chipper/pkg/vtt"
+	sr "github.com/kercre123/chipper/pkg/wirepod/speechrequest"
+	ttr "github.com/kercre123/chipper/pkg/wirepod/ttr"
+)
+
+func (s *Server) ProcessIntent(req *vtt.IntentRequest) (*vtt.IntentResponse, error) {
+	sr.BotNum = sr.BotNum + 1
+	var successMatched bool
+	speechReq := sr.ReqToSpeechRequest(req)
+	transcribedText, err := sttHandler(speechReq)
+	if err != nil {
+		sr.BotNum = sr.BotNum - 1
+		ttr.IntentPass(req, "intent_system_noaudio", "voice processing error", map[string]string{"error": err.Error()}, true, speechReq.BotNum)
+		return nil, nil
+	}
+	successMatched = ttr.ProcessTextAll(req, transcribedText, matchListList, intentsList, speechReq.IsOpus, speechReq.BotNum)
+	if !successMatched {
+		logger.Println("No intent was matched.")
+		sr.BotNum = sr.BotNum - 1
+		ttr.IntentPass(req, "intent_system_noaudio", transcribedText, map[string]string{"": ""}, false, speechReq.BotNum)
+		return nil, nil
+	}
+	sr.BotNum = sr.BotNum - 1
+	logger.Println("Bot " + strconv.Itoa(speechReq.BotNum) + " request served.")
+	return nil, nil
+}
